20221225 completed atmosphere opacity

+++ atmo_shader_vert.glsl
@@ -6,8 +6,8 @@ uniform vec3 sun_position;
 uniform float planet_radius;
 uniform float atmo_radius;
 
-uniform int view_path_samples;
-uniform int light_path_samples;
+uniform float view_path_samples;
+uniform float light_path_samples;
 
 // // = object.matrixWorld
 // uniform modelMatrix;
